requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# ClickHouse
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Django (for backoffice)
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2